+++ mslice/widgets/cut/cut.py
@@ -11,7 +11,7 @@ from qtpy.QtWidgets import QWidget
 from mslice.models.cut.mantid_cut_algorithm import MantidCutAlgorithm
 from mslice.models.cut.matplotlib_cut_plotter import MatplotlibCutPlotter
 from mslice.presenters.cut_presenter import CutPresenter
-from mslice.util.load_ui import load_ui
+from mslice.util.qt import load_ui
 from mslice.views.cut_view import CutView
 from .command import Command
