+++ mslice/widgets/slice/slice.py
@@ -11,7 +11,7 @@ from qtpy.QtWidgets import QWidget
 from mslice.models.slice.mantid_slice_algorithm import MantidSliceAlgorithm
 from mslice.models.slice.matplotlib_slice_plotter import MatplotlibSlicePlotter
 from mslice.presenters.slice_plotter_presenter import SlicePlotterPresenter
-from mslice.util.load_ui import load_ui
+from mslice.util.qt import load_ui
 from mslice.views.slice_plotter_view import SlicePlotterView
 from .command import Command
